convert to Path.cs to immutable struct
